requirements.txt
Flask==3.0.0
gunicorn==21.2.0
reportlab==4.0.8
google-auth==2.23.4
requests==2.31.0
google-api-python-client==2.108.0